refactor(query-builder): move key-value operation utilities to a core module
